requirements.txt
Package             Version
------------------- -------
asgiref             3.2.7  
Django              3.0.4  
djangorestframework 3.11.0 
pip                 19.0.3 
PyJWT               1.7.1  
pytz                2019.3 
setuptools          40.8.0 
sqlparse            0.3.1  
